+++ src/main/resources/db/migration/20210609104652976__add_constraint_to_facility_type_table.sql
@@ -1,0 +1,6 @@
+--
+-- Name: facility_types_uk_name; Type: CONSTRAINT; Schema: referencedata; Owner: postgres; Tablespace:
+--
+
+ALTER TABLE ONLY facility_types
+    ADD CONSTRAINT facility_types_uk_name UNIQUE (name);