+++ src/main/resources/db/migration/20210609130104095__rename_facility_type_unique_name_constraint.sql
@@ -1,0 +1,4 @@
+ALTER TABLE facility_types
+    DROP CONSTRAINT facility_types_uk_name;
+
+CREATE UNIQUE INDEX unq_facility_type_name ON facility_types (LOWER(name));